backend/requirements.txt
# Core FastAPI
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
email-validator

# Database
motor
pymongo

# Authentication
PyJWT
bcrypt

# HTTP
httpx
aiohttp

# OpenAI
openai

# Utilities
python-multipart
python-dateutil
